Improve Contributing Guidelines
I tried to keep the best of both worlds with the original and newly
proposed contribution doc and this one contains a terse version for
those already experienced with git and a long version for those that
might need it. I also removed the "Things to know" section and linked
it to the website instead, which has the same info.